% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{plotTimeSeries}
\usage{
plotTimeSeries(country, dateFrom, dateTo)
}
\arguments{
\item{country}{Country Name e.g. "Iran"}

\item{dateFrom}{Starting date e.g. "2020-05-11"}

\item{dateTo}{Ending Date e.g. "2020-08-05"}
}
\value{
Times series plot
}
\description{
Plot death and confirmed rates for a country between two dates

Dates must be in "YYYY-MM-DD" format
}
\examples{
plotTimeSeries("Iran", "2020-05-11", "2020-08-05")

}
