% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWorld.R
\name{plotWorld}
\alias{plotWorld}
\title{plotWorld}
\usage{
plotWorld(plotType, date)
}
\arguments{
\item{plotType}{Type of plot: "Deaths" or "Confirmed"}

\item{date}{e.g "2020-05-11"}
}
\value{
World map plot
}
\description{
Plot daily cases in world map.
}
\examples{
plotWorld("Confirmed", "2020-05-11")

}
